+++ CMakeLists.txt
@@ -204,9 +204,11 @@ set(SOURCES
 		src/cmd/track.cpp
 		src/cmd/hire.cpp
 		src/cmd/telegram.cpp
+		src/cmd.wiz/stat.cpp
 )
 
 set(HEADERS
+		src/cmd.wiz/stat.h
 		src/cmd/telegram.h
 		src/cmd/hire.h
 		src/cmd/telegram.h
