+++ src/fightsystem/fight_hit.hpp
@@ -89,6 +89,8 @@ private:
 	flags_t m_flags;
 };
 
+int compute_armor_class(CHAR_DATA * ch);
+
 int check_agro_follower(CHAR_DATA * ch, CHAR_DATA * victim);
 void set_battle_pos(CHAR_DATA * ch);
